Update User, implement Add Custom Food and Exercise

+++ database/migrations/2022_06_06_031647_create_processes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->bigInteger('user_id')->unsigned();
             $table->integer('height');
-            $table->integer('starting_weight');
+            $table->float('starting_weight');
             $table->float('current_weight');
             $table->integer('goal_weight');
             $table->string('weekly_goal');
